pyproject.toml
[tool.poetry]
name = "kibune"
version = "0.1.0"
description = ""
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aioredis = "1.3.1"
arq = "0.22"
dict-hash = "^1.1.26"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
httpx = "^0.23.0"
importlib-metadata = "^4.12.0"
lark-parser = "^0.12.0"
loguru = "^0.6.0"
mysqlclient = "^2.1.1"
pydantic = "^1.9.1"
pyhumps = "^3.7.2"
pysigma = "^0.6.5"
python-ulid = "^1.1.0"
PyYAML = "^6.0"
regex = "^2022.7.9"
SQLAlchemy = ">=1.4.17,<=1.4.35"
sqlmodel = "^0.0.6"
ujson = "^5.4.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
autopep8 = "^1.6.0"
black = "^22.3.0"
coveralls = "^3.3.1"
isort = "^5.10.1"
mypy = "^0.961"
pre-commit = "^2.20.0"
pytest = "^6.2.5"
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"
pytest-httpserver = "^1.0.4"
pytest-mock = "^3.8.2"
pytest-randomly = "^3.12.0"
pytest-sugar = "^0.9.5"
pyupgrade = "^2.37.1"
types-PyYAML = "^6.0.10"
diagrams = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
